refactor: separate StoryOperations::Update into its own file
